+++ Editor/Scripts/Model/Orientation.cs
@@ -1,0 +1,8 @@
+namespace dev.klebersilva.tools.bitmapfontcreator
+{
+	internal enum Orientation
+	{
+		Horizontal,
+		Vertical,
+	}
+}