+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0]
+
+### Added
+
+-  Default button
+
+### Changed
+
+-  Monospaced fonts now respects custom character properties
+-  Keeping changes at runtime
+
+---
+
 ## [1.4.0]
 
 ### Added
@@ -119,6 +132,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial version
 
+[1.5.0]: https://github.com/kleber-swf/unity-bitmap-font-creator/releases/tag/1.5.0
 [1.4.0]: https://github.com/kleber-swf/unity-bitmap-font-creator/releases/tag/1.4.0
 [1.3.0]: https://github.com/kleber-swf/unity-bitmap-font-creator/releases/tag/1.3.0
 [1.2.1]: https://github.com/kleber-swf/unity-bitmap-font-creator/releases/tag/1.2.1